M src/server/j2se/com/sun/sgs/impl/service/data/store/db/je/JeEnvironment.java
- Modify constructor to add missing long cast.

M src/shared/j2se/com/sun/sgs/impl/sharedutil/Objects.java

- Make this class final, to make findbugs happy.
- Modify safeToString to add some missing doc, and to only catch
  RuntimeExceptions, since there are no declared exceptions being
  thrown.
- Modify fastToString to add some missing doc.

+++ src/server/j2se/com/sun/sgs/impl/service/data/store/db/je/JeEnvironment.java
@@ -286,7 +286,7 @@ public class JeEnvironment implements DbEnvironment {
 	    TransactionCoordinator.TXN_TIMEOUT_PROPERTY, -1);
 	long defaultLockTimeout = (txnTimeout < 1)
 	    ? DEFAULT_LOCK_TIMEOUT
-	    : txnTimeout / DEFAULT_LOCK_TIMEOUT_PROPORTION;
+	    : (long) (txnTimeout / DEFAULT_LOCK_TIMEOUT_PROPORTION);
 	/* Avoid underflow */
 	if (defaultLockTimeout < 1) {
 	    defaultLockTimeout = 1;
